+++ EmployeeSkillsKnowledgeBase/src/main/java/com/rsi/esk/util/NumberUtils.java
@@ -1,7 +1,7 @@
 package com.rsi.esk.util;
 
 public class NumberUtils {
-    public static Boolean hasInteger(Integer value) {
+    public static Boolean hasLong(Long value) {
         if (value == null) {
             return false;
         }
